Create Final genomic analysis
The analysis of an genomic mapping aswell as a transcript analysis.
As our pipeline does not support genomic analysis, I have made a fasta file with the forward strand and the reverse strand in reverse complement. I can later change the position numbers on the reverse complement data.
The transcript file is made from the manually annotated files that Lars has sent.